Plug some GCC fortify warnings

+++ file_subs.c
@@ -1158,7 +1158,7 @@ set_crc(ARCHD *arcn, int fd)
 	int i;
 	int res;
 	off_t cpcnt = 0L;
-	u_long size;
+	size_t size;
 	u_int32_t crc = 0;
 	char tbuf[FILEBLK];
 	struct stat sb;
@@ -1171,8 +1171,8 @@ set_crc(ARCHD *arcn, int fd)
 		return(0);
 	}
 
-	if ((size = (u_long)arcn->sb.st_blksize) > (u_long)sizeof(tbuf))
-		size = (u_long)sizeof(tbuf);
+	if ((size = (size_t)arcn->sb.st_blksize) > sizeof(tbuf))
+		size = sizeof(tbuf);
 
 	/*
 	 * read all the bytes we think that there are in the file. If the user
